Move MenuItem within Menu, use MVVM model.
